requirements.txt
python-dotenv
streamlit
groq